requirements.txt
wxPython==4.2.0
pandas==1.4.2
python-dotenv==0.21.0
selenium==4.5.0
xlsxwriter==3.0.3